requirements.txt
python-dotenv
djangorestframework
django
psycopg2
django-filter